+++ src/dashboard.html
@@ -39,4 +39,4 @@
     <div id="bottom-panel"></div>
   </div>
 </div>
-<script src="./ui/js/dashboard.js"></script>
+<script type="module" src="./ui/js/dashboard.js"></script>
